requirements.txt
absl-py==1.2.0
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
astunparse==1.6.3
attrs==22.2.0
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
distlib==0.3.6
faiss-cpu==1.7.3
fastapi==0.95.0
filelock==3.9.0
Flask==2.2.2
flatbuffers==2.0.7
frozenlist==1.3.3
ftfy==6.1.1
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.49.0
h11==0.14.0
h5py==3.7.0
httptools==0.5.0
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.10.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
mpmath==1.3.0
msgpack==1.0.4
multitasking==0.0.11
networkx==3.1
numpy==1.23.3
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
Pillow==9.5.0
platformdirs==3.0.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.7
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.5
PyYAML==6.0
ray==2.2.0
regex==2023.3.23
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-macos==2.10.0
tensorflow-metal==0.6.0
termcolor==2.0.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.21.1
uvloop==0.17.0
virtualenv==20.19.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.1
Werkzeug==2.2.2
wrapt==1.14.1
yfinance==0.1.79
zipp==3.8.1
